+++ Source/Core/Core/PowerPC/Jit64/GPRRegCache.h
@@ -1,0 +1,19 @@
+// Copyright 2016 Dolphin Emulator Project
+// Licensed under GPLv2+
+// Refer to the license.txt file included.
+
+#pragma once
+
+#include "Core/PowerPC/Jit64/JitRegCache.h"
+
+class GPRRegCache final : public RegCache
+{
+public:
+  void StoreRegister(size_t preg, const Gen::OpArg& new_loc) override;
+  void LoadRegister(size_t preg, Gen::X64Reg new_loc) override;
+  Gen::OpArg GetDefaultLocation(size_t reg) const override;
+  const Gen::X64Reg* GetAllocationOrder(size_t* count) override;
+  void SetImmediate32(size_t preg, u32 imm_value, bool dirty = true);
+  BitSet32 GetRegUtilization() override;
+  BitSet32 CountRegsIn(size_t preg, u32 lookahead) override;
+};
